Moved everything to two screens instead of one file + routing + cleaning code
