+++ lib/screens/todo_detail_screen.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class TodoDetailScreen extends StatelessWidget {
+  static const routeName = '/todo-detail';
+
+  const TodoDetailScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Text('Sexi time');
+  }
+}
